Fix template error for scope_wrap. Update notification's time to be using ImGui's delta time.

+++ imcxx/include/imgui/imcxx/misc/notification.hpp
@@ -72,7 +72,7 @@ namespace imcxx::misc
 		/// <summary>
 		/// notification's duration in milli-seconds
 		/// </summary>
-		uint64_t Duration;
+		float Duration;
 		/// <summary>
 		/// notification's background/foreground color
 		/// </summary>
